Fold stack_live.go into stack_config.go
